--- web/html/src/manager/content-management/shared/components/down-arrow/down-arrow.js
@@ -1,18 +1,0 @@
-// @flow
-import * as React from 'react';
-
-const DownArrow = () => {
-  return (
-    <div style={{
-      fontSize: "26px",
-      textAlign: "center",
-      color: "#369",
-      width: "26px",
-      margin: "auto",
-      transform: "rotate(90deg)"
-    }}> ➙
-    </div>
-  );
-};
-
-export default DownArrow;
